chore: move alt -> index
